+++ src/lib.rs
@@ -21,9 +21,9 @@ pub use instance::{Instance, RegisterInstance};
 pub use registry::Registry;
 pub use resolve::{Resolve, ResolveOwned, ResolveRef};
 pub use types::{
-    Enum, ExpandObject, GetInputTypeRef, GetOutputTypeRef, GraphqlType, InputObject, InputType,
-    Interface, InterfaceMark, Mutation, Object, OutputType, ParentType, Register, RegisterFns,
-    Scalar, Union,
+    Enum, ExpandObject, GetInputTypeRef, GetOutputTypeRef, InputObject, InputType, Interface,
+    InterfaceMark, Mutation, Object, OutputType, ParentType, Register, RegisterFns, Scalar,
+    TypeName, Union,
 };
 pub use upload::Upload;