+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - remove `MARK` from `Interface` trait
 - use function instead constant for type names
+- rename `GraphqlType` to `TypeName`
 
 ## [0.5.4] - 2023-01-30